+++ src/commands/uploademoji.js
@@ -1,9 +1,9 @@
-const { Permissions } = require('discord.js');
+const { Permissions, MessageEmbed } = require('discord.js');
 const { SlashCommandBuilder } = require('@discordjs/builders');
 const axios = require('axios');
 const fs = require('fs');
 const { exec } = require('child_process');
-const { sendErrorFeedback } = require('../helpers/utilities');
+const { sendErrorFeedback, mediaLinks } = require('../helpers/utilities');
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -95,7 +95,10 @@ module.exports = {
 						interaction.guild.emojis
 							.create(path, name)
 							.then(emoji => {
-								return interaction.editReply({ content: `Added ${emoji} to server!` });
+								const embed = new MessageEmbed()
+									.setTitle(`Added ${emoji} to server!`)
+									.setDescription(`If you've enjoyed this bot so far, please consider voting for it.\nIt helps the bot grow. 🙂\n${mediaLinks}`);
+								return interaction.editReply({ embeds: [embed] });
 							})
 							.catch(e => {
 								console.error(e);
@@ -113,7 +116,10 @@ module.exports = {
 				interaction.guild.emojis
 					.create(buffer, name)
 					.then(emoji => {
-						return interaction.editReply({ content: `Added ${emoji} to server!` });
+						const embed = new MessageEmbed()
+							.setTitle(`Added ${emoji} to server!`)
+							.setDescription(`If you've enjoyed this bot so far, please consider voting for it.\nIt helps the bot grow. 🙂\n${mediaLinks}`);
+						return interaction.editReply({ embeds: [embed] });
 					})
 					.catch(e => {
 						return interaction.editReply({ content: `Emoji creation failed!\n${e.message}` });